Jm/branch2

+++ be/src/main/java/com/example/be/service/member/MemberService.java
@@ -2,15 +2,27 @@ package com.example.be.service.member;
 
 import com.example.be.dto.member.Member;
 import com.example.be.dto.member.MemberEdit;
+import com.example.be.dto.member.MemberPicture;
 import com.example.be.mapper.member.MemberMapper;
+import com.example.be.mapper.tour.TourMapper;
+import com.example.be.service.tour.TourService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.oauth2.jwt.JwtClaimsSet;
 import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.jwt.JwtEncoderParameters;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+import software.amazon.awssdk.core.sync.RequestBody;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.ObjectCannedACL;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
+import java.io.IOException;
 import java.time.Instant;
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -18,10 +30,39 @@ import java.util.List;
 @RequiredArgsConstructor
 public class MemberService {
   final MemberMapper mapper;
+  final S3Client s3;
   final JwtEncoder jwtEncoder;
 
-  public boolean add(Member member) {
+  private final TourService tourService;
+  private final TourMapper tourMapper;
+
+  @Value("${image.src.prefix}")
+  String imageSrcPrefix;
+
+  @Value("${bucket.name}")
+  String bucketName;
+
+  public boolean add(Member member, MultipartFile[] files) {
     int cnt = mapper.insert(member);
+
+    if (files != null && files.length > 0) {
+      for (MultipartFile file : files) {
+        mapper.updatePicture(member.getEmail(), file.getOriginalFilename());
+
+        String objectKey = "teamPrj1126/member/" + member.getEmail() + "/" + file.getOriginalFilename();
+        PutObjectRequest por = PutObjectRequest.builder()
+                .bucket(bucketName)
+                .key(objectKey)
+                .acl(ObjectCannedACL.PUBLIC_READ)
+                .build();
+        try {
+          s3.putObject(por, RequestBody.fromInputStream(file.getInputStream(), file.getSize()));
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
+      }
+    }
+
     return cnt == 1;
   }
 
@@ -34,29 +75,82 @@ public class MemberService {
   }
 
   public Member get(String email) {
-    return mapper.selectByEmail(email);
+    Member member = mapper.selectByEmail(email);
+    String profileName = mapper.selectPictureByEmail(email);
+
+    if (profileName != null) {
+      MemberPicture PicSrc = new MemberPicture(
+              profileName, imageSrcPrefix + "/member/" + email + "/" + profileName);
+      member.setProfile(List.of(PicSrc));
+    } else {
+      member.setProfile(Collections.emptyList());
+    }
+
+    return member;
   }
 
   public boolean remove(Member member) {
     int cnt = 0;
 
     Member db = mapper.selectByEmail(member.getEmail());
-    if (db != null) {
-      if (db.getPassword().equals(member.getPassword())) {
-        cnt = mapper.deleteByEmail(member.getEmail());
+    if (db != null && db.getPassword().equals(member.getPassword())) {
+
+      List<Integer> tourBoards = tourMapper.selectByPartner(db.getNickname());
+      for (Integer tourId : tourBoards) {
+        tourService.delete(tourId);
       }
+
+      String profile = mapper.selectPictureByEmail(db.getEmail());
+      String key = "teamPrj1126/member/" + member.getEmail() + "/" + profile;
+      DeleteObjectRequest dor = DeleteObjectRequest.builder()
+              .bucket(bucketName)
+              .key(key)
+              .build();
+      s3.deleteObject(dor);
+
+      cnt = mapper.deleteByEmail(member.getEmail());
+
     }
+
     return cnt == 1;
   }
 
-  public boolean update(MemberEdit member) {
+  public boolean update(MemberEdit member, MultipartFile uploadFiles) {
     int cnt = 0;
-    Member db = mapper.selectByEmail(member.getEmail());
-    if (db != null) {
-      if (db.getPassword().equals(member.getOldPassword())) {
-        cnt = mapper.update(member);
 
+    //회원 조회 & 비번 대조
+    Member db = mapper.selectByEmail(member.getEmail());
+    if (db != null && db.getPassword().equals(member.getOldPassword())) {
+      if (uploadFiles != null) {
+        //s3기존 이미지 삭제
+        String oldPicture = mapper.selectPictureByEmail(db.getEmail());
+        String key = "teamPrj1126/member/" + member.getEmail() + "/" + oldPicture;
+        DeleteObjectRequest dor = DeleteObjectRequest.builder()
+                .bucket(bucketName)
+                .key(key)
+                .build();
+        s3.deleteObject(dor);
+
+
+        //member 테이블에 이미지 정보 수정
+        mapper.updatePicture(member.getEmail(), uploadFiles.getOriginalFilename());
+
+        //s3 이미지 업로드
+        String objectKey = "teamPrj1126/member/" + member.getEmail() + "/" + uploadFiles.getOriginalFilename();
+        PutObjectRequest por = PutObjectRequest.builder()
+                .bucket(bucketName)
+                .key(objectKey)
+                .acl(ObjectCannedACL.PUBLIC_READ)
+                .build();
+
+        try {
+          s3.putObject(por, RequestBody.fromInputStream(uploadFiles.getInputStream(), uploadFiles.getSize()));
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
       }
+
+      cnt = mapper.update(member);
     }
     return cnt == 1;
   }
